Rename Button code file
