+++ lib/views/daycare_index.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:my_view_app_001/Components/button_default_a.dart';
+import 'package:my_view_app_001/Components/button_container_element.dart';
 import 'package:my_view_app_001/views/daycare_agreement_user.dart';
 
 class DaycareIndexView extends StatelessWidget {
